analytics-service/api/grpc_servicer.py
"""
Analytics Service gRPC Servicer
Implements the AnalyticsService gRPC interface
"""

import grpc
import logging

logger = logging.getLogger(__name__)


class AnalyticsServiceServicer:
    """
    gRPC servicer for Analytics Service
    Implements analytics_pb2_grpc.AnalyticsServiceServicer
    """
    
    def __init__(self, business_service):
        """Initialize servicer with business service"""
        self.business_service = business_service
        logger.info("AnalyticsServiceServicer initialized")
    
    async def QueryMarketData(self, request, context):
        """
        Query raw market data points
        
        Args:
            request: QueryRequest from proto
            context: gRPC context
            
        Returns:
            QueryResponse
        """
        # TODO: Implement QueryMarketData
        pass
    
    async def GetCandles(self, request, context):
        """
        Get aggregated candles (OHLCV)
        
        Args:
            request: CandleRequest from proto
            context: gRPC context
            
        Returns:
            CandleResponse
        """
        # TODO: Implement GetCandles
        pass
    
    async def GetAggregatedMetrics(self, request, context):
        """
        Get calculated metrics (volatility, averages, etc.)
        
        Args:
            request: AggregationRequest from proto
            context: gRPC context
            
        Returns:
            AggregationResponse
        """
        # TODO: Implement GetAggregatedMetrics
        pass
    
    async def StreamRealTimeData(self, request, context):
        """
        Stream real-time market data
        
        Args:
            request: StreamRequest from proto
            context: gRPC context
            
        Yields:
            MarketDataEvent
        """
        # TODO: Implement StreamRealTimeData
        pass


# TODO: Implement error handling
# TODO: Add request validation
# TODO: Add authentication check

